requirements.txt
numpy==1.13.3
torch==0.4.0
scipy==1.0.0
librosa==0.5.1
tqdm==4.19.4
config==0.3.9
scikit_learn==0.19.1
PyYAML==3.12
